Move generational version of update refs to generational heap
